+++ CHANGELOG.md
@@ -4,6 +4,16 @@ Todos los cambios notables en este proyecto se documentarán en este archivo.
 
 El formato se basa (en su mayor parte) en [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) y este proyecto se adhiere al [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Próximos Cambios] "¿Aliens?"
+
+### Añadido
+
+- Se agrega la función `CreateBlankEntity` para crear una entidad que solo tiene el componente `C_Transform`.  
+
+### Eliminado
+
+- Se elimina la función `DefineComponents` debido al método de gestión de componentes de Flecs.  
+
 ## [1.1] "Explorador" - 2025-02-02
 
 ### Añadido